pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Ryley Oroku", email = "greeeenmango89@gmail.com"},
]
dependencies = [
    "discord.py>=1.7.3",
    "ffmpeg-python>=0.2.0",
]
requires-python = ">=3.7"
license = {text = "MIT"}

[tool.pdm]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
